backend/requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
PyPDF2==2.11.1
python-docx==0.8.11
pytesseract==0.3.10
Pillow==9.5.0
spacy==3.5.3
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0